% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allDuplicated.R
\name{allDuplicated}
\alias{allDuplicated}
\title{Gets all duplicated values}
\usage{
allDuplicated(x)
}
\arguments{
\item{x}{An object on which the \code{duplicated} function could be applied}
}
\description{
Returns every appearance of a duplicated value
}
\examples{
allDuplicated(c(1, 2, 3, 1, 1))

}
