% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigSummary.R
\name{bigSummary}
\alias{bigSummary}
\title{Generates an extensive summary of a numerical variable}
\usage{
bigSummary(dvar)
}
\arguments{
\item{dvar}{The column of the data frame to investigate}
}
\description{
This prints the name of the variable, a summary and a density plot
}
