% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCWFrequencies.R
\name{calcCWFrequencies}
\alias{calcCWFrequencies}
\title{Calculates geocode to geocode match}
\usage{
calcCWFrequencies(dat, geocode1, geocode2)
}
\arguments{
\item{dat}{A crosswalk file from geocorr}

\item{geocode1}{Name of the variable containing the first geocode}

\item{geocode2}{Name of the variable containing the second geocode}
}
\description{
Calculates how many geocodes of one level match to one geocode of the other level, gives out the number of different geocodes, the maximum number of geocodes matching to one geocode and a frequency plot.
}
