% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cDist.R
\name{cDist}
\alias{cDist}
\title{Looks up distances from a given table}
\usage{
cDist(a, b, dists)
}
\arguments{
\item{a, b}{Location vectors}

\item{dists}{Table containing distances with geoids as column and row names (e.g. generated by \code{geoDistCalc()})}
}
\value{
Gives a vector containing the distance for each pairing
}
\description{
Takes two vectors of locations and a distance table containing location ids and looks up the distance.
}
