% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covTex.R
\name{covTex}
\alias{covTex}
\title{A covariance LaTex generator}
\usage{
covTex(x, y)
}
\arguments{
\item{x}{The first vector}

\item{y}{The second vector}
}
\description{
This generates the covariance formula in LaTex code for given values. Note that both vectors have to be the same length
}
\examples{
covTex(c(1,2,3), c(2,3,4))

}
