% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{create a package project}
\usage{
create_package(path, author)
}
\arguments{
\item{path}{project path.}

\item{author}{used for MIT license.}
}
\value{
Rpackages project
}
\description{
A function to create a package project combining common process in my own
package development
}
\examples{
# ADD_EXAMPLES_HERE
}
