% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwAutomatic.R
\name{cwAutomatic}
\alias{cwAutomatic}
\title{Crosswalk function}
\usage{
cwAutomatic(df, cw, varMerge, varName)
}
\arguments{
\item{df}{A data frame with one column with the same name as one column in the crosswalk}

\item{cw}{A crosswalk as data.frame with two variables, output of giveCw should work}

\item{varMerge}{String giving the name of the variable that should be merged}

\item{varName}{String giving the desired name for the output variable}
}
\description{
A function that crosswalks and names the outcoming variable
}
