\name{Doraemon}
\alias{Doraemon}
\title{Cartoon of Doraemon}
\description{This function draws the cartoon of "Doraemon". Note: the eyes of Doraemon will keep tracking your mouse, and the small bell will jingle when your mouse move on it.}
\usage{Doraemon()}
\author{Huidong Tian}
\examples{Doraemon()}
