% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floorCeil.R
\name{floorCeil}
\alias{floorCeil}
\title{Floor or ceiling with number decimals option}
\usage{
floorCeil(x, n = 2, FUN = floor, cutLZero = TRUE)
}
\arguments{
\item{x}{Vector of numbers to round}

\item{n}{Number of numbers after decimal point}

\item{FUN}{Function to use on numbers, for ceiling use \code{ceiling}}

\item{cutLZero}{Whether zero before the decimal point should be left out.}
}
\description{
Floor or ceiling of a number to the number of decimals given and the option to leave out the zero before the decimal point.
}
\examples{
floorCeil(0.4592)

}
