% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDistCalc.R
\name{geoDistCalc}
\alias{geoDistCalc}
\title{Calculates distances between all given shapes}
\usage{
geoDistCalc(sf, vars = NULL)
}
\arguments{
\item{sf}{Shapefile to calculate distances on}

\item{vars}{Name(s) of the variables that should be simulated}
}
\value{
Gives a matrix containing all distances between all pairings. If variables are simulated, gives a list containing the distances and a data frame with the variables together with the GEOID.
}
\description{
Calculates direct distances between every pairing of the given shapes. If necessary, simulates random variables for each geographic entity and gives them in a data frame.
}
