% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDateTz.R
\name{getDateTz}
\alias{getDateTz}
\title{!!Preliminary!! function to extract times from the GPIPP!!}
\usage{
getDateTz(x)
}
\arguments{
\item{x}{The date string}
}
\description{
First steps to extract times and time zones from older answers to the GPIPP since it has different date formats
}
