% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giveCw.R
\name{giveCw}
\alias{giveCw}
\title{Crosswalk prepare function}
\usage{
giveCw(cw, from, to, by, ftb = c(from, to, by))
}
\arguments{
\item{cw}{A crosswalk as data.frame}

\item{from}{Variable from which the crosswalk should happen}

\item{to}{Variable to which crosswalk should happen}

\item{by}{Variable for selecting which case in case of duplicates}

\item{ftb}{Variables can also be given as vector}
}
\description{
A function that prepares a crosswalk containing no missings and no duplicates based on a given variable
}
