% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giveIntervals.R
\name{giveIntervals}
\alias{giveIntervals}
\title{Calculates a list with interval information based on lower bounds}
\usage{
giveIntervals(lBounds)
}
\arguments{
\item{lBounds}{A vector containing all lower bounds and the upper bound of the last interval}
}
\description{
This gives out a list containing the lower bounds, weights (i.e., mid points of the intervals), and widths of the intervals. Note that the output is one shorter than the input since the last lower bound has no corresponding upper bound.
}
\examples{
giveIntervals(c(1, 2, 5, 7))

}
