% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givePrice.R
\name{givePrice}
\alias{givePrice}
\title{A function that computes the better buying option for cookie clicker buildings}
\usage{
givePrice(p1, cps1, p2, cps2)
}
\arguments{
\item{p1}{Price of the first building, doesn't need to be the actual price, can also be a share of the price as long as it's consistent to the second building.}

\item{cps1}{Cookie production of the first building, doesn't need to be the actual production, can also be a share of the production as long as it's consistent to the second building.}

\item{p2}{Price of the second building, doesn't need to be the actual price, can also be a share of the price as long as it's consistent to the first building.}

\item{cps2}{Cookie production of the second building, doesn't need to be the actual production, can also be a share of the production as long as it's consistent to the first building.}
}
\description{
This gives the better option of two buildings based on current price and cookie production
}
\examples{
givePrice(100, 100, 30, 100)

}
