% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpMedian.R
\name{interpMedian}
\alias{interpMedian}
\title{Interpolates median for ranges based on frequencies}
\usage{
interpMedian(x, lBounds, widths)
}
\arguments{
\item{x}{A vector containing the frequencies}

\item{lBounds}{A vector containing the lower bounds of the ranges}

\item{widths}{A vector containing the widths of the ranges}
}
\description{
This interpolates the median if given frequencies, lower bounds of ranges, and widths of ranges
}
