% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipumsAge.R
\name{ipumsAge}
\alias{ipumsAge}
\title{A function that calculates median and mean age for IPUMS data}
\usage{
ipumsAge(df, code)
}
\arguments{
\item{df}{The data frame to calculate for}

\item{code}{The three letter code for age}
}
\description{
Calculates mean and interpolated median age for IPUMS data sets based on frequency tables.
}
