% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipumsVars.R
\name{ipumsVars}
\alias{ipumsVars}
\title{IPUMS variable codes generator}
\usage{
ipumsVars(code, n, type = "estimate")
}
\arguments{
\item{code}{The 3 letter code of the IPUMS variable, the fourth letter indicates whether it's estimate (E) or margin of error (M)}

\item{n}{The number of the IPUMS variable that comes after the letter code. It's sufficient to give the number without the leading zeros.}

\item{type}{The type of value wanted: "estimate" for Estimate, "margin" for Margin of Error, misspecify for further information on possible values., if NULL, only \code{code} is used.}
}
\description{
This generates IPUMS variable codes based on letter/number code(s) and number(s)
}
\examples{
ipumsVars(code = "ABC", n = 1:3)
ipumsVars(c("ABC", "DEF"), 1, type = "margin")
try(ipumsVars("ABC", 1, "wrong"))

}
