% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mBib.R
\name{mBib}
\alias{mBib}
\title{Adds citations for packages to existing .bib file}
\usage{
mBib(path, pckgs)
}
\arguments{
\item{path}{String containing the path to the .bib file}

\item{pckgs}{Vector of Strings giving the used packages}
}
\description{
Adds citations for packages to existing .bib file
}
