% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFaTable.R
\name{mFaTable}
\alias{mFaTable}
\title{Generates a tidy looking table from factor analysis results}
\usage{
mFaTable(fa, cutoff, nRound = 3)
}
\arguments{
\item{fa}{Results from factor analysis}

\item{cutoff}{Value below which results should be left out of the reduced table}

\item{nRound}{number of digits for rounding}
}
\description{
Generates a tidy looking table from factor analysis results
}
