% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moTheme.R
\docType{data}
\name{moTheme}
\alias{moTheme}
\title{My Theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 4.
}
\usage{
moTheme
}
\description{
A very clean theme for ggplot2 with no axes and no background
}
\keyword{datasets}
