% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mTex.R
\name{mTex}
\alias{mTex}
\title{A mean LaTex generator}
\usage{
mTex(x, letter = "x")
}
\arguments{
\item{x}{A vector}

\item{letter}{The letter to use inside the formula, defaults to x}
}
\description{
This generates the mean formula in LaTex code for given values.
}
\examples{
mTex(c(1,2,3))
mTex(c(2,3,4), letter = "y")

}
