% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predTex.R
\name{predTex}
\alias{predTex}
\title{A linear prediction LaTex generator}
\usage{
predTex(x, y, px)
}
\arguments{
\item{x}{The first vector}

\item{y}{The second vector}

\item{px}{The value of x to predict y by}
}
\description{
This generates the linear prediction formula in LaTex code for given values. Note that both vectors have to be the same length
}
\examples{
predTex(c(1,2,3), c(2,3,4), 1)

}
