% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.crosswalk.R
\name{read.crosswalk}
\alias{read.crosswalk}
\title{Reads a crosswalk file}
\usage{
read.crosswalk(path)
}
\arguments{
\item{path}{Path of the crosswalk}
}
\description{
Convenience function for crosswalks from geocorr
}
