% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redGreenConverter.R
\name{redGreenConverter}
\alias{redGreenConverter}
\title{Converts red-green scales to red-blue scales}
\usage{
redGreenConverter(path = file.choose(), output = path)
}
\arguments{
\item{path}{The path where the image can be found}

\item{output}{The path where the converted image should be stored.}
}
\description{
Function allows easy conversion to red-green blind friendly color scale.
}
