% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_template.R
\name{rmd_template}
\alias{rmd_template}
\title{Rmd file template}
\usage{
rmd_template(template, dir = getwd())
}
\arguments{
\item{template}{Rmd template.}

\item{dir}{where to create template file.}
}
\value{
RETURN_DESCRIPTION
}
\description{
A Rmd templates file selector
}
\examples{
# ADD_EXAMPLES_HERE
}
