% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.p.R
\name{sd.p}
\alias{sd.p}
\title{Calculate population standard deviation}
\usage{
sd.p(x, na.rm = T)
}
\arguments{
\item{x}{Vector of numbers to calculate population standard deviation for}

\item{na.rm}{Whether \code{NA}s should be removed}
}
\description{
Calculate population standard deviation
}
\examples{
sd.p(1:10)

}
