% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDat.R
\name{simDat}
\alias{simDat}
\title{Function to simulate simple data}
\usage{
simDat(ids, vars, dichVars = NULL)
}
\arguments{
\item{ids}{Vector of ids to identify each row}

\item{vars}{Vector of variable names for continuous normally distributed
variables}

\item{dichVars}{Vector of variable names for dichotomous uniformly
distributed variables}
}
\description{
Generates a dataframe with one row for each id, normally distributed
variables for 'vars', and uniform dichotomous variables for 'dichVars'
}
