% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_mult.R
\name{tab_mult}
\alias{tab_mult}
\title{Crea una tabla de multiplicar}
\usage{
tab_mult(n)
}
\arguments{
\item{n}{numeric}
}
\value{
matrix
}
\description{
Crea una tabla de multiplicar
}
\examples{
tab_mult(10)
}
