% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRFun.R
\name{updateRFun}
\alias{updateRFun}
\title{Convenience function for updating RFun}
\usage{
updateRFun(...)
}
\arguments{
\item{...}{Arguments to be passed on to \code{install_github()}}
}
\description{
Unloads RFun, installs the newest version and attaches it again
}
