% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTex.R
\name{varTex}
\alias{varTex}
\title{A variance LaTex generator}
\usage{
varTex(x, letter = "x")
}
\arguments{
\item{x}{A vector}

\item{letter}{The letter to use inside the formula, defaults to x}
}
\description{
This generates the variance formula in LaTex code for given values.
}
\examples{
varTex(c(1,2,3))
varTex(c(2,3,4), letter = "y")

}
