% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_roxygen.R
\name{write_roxygen}
\alias{write_roxygen}
\title{add roxygen sekleton for function}
\usage{
write_roxygen(fun)
}
\arguments{
\item{fun}{funtion name.}
}
\value{
roxygen sekleton.
}
\description{
use roxygen sekleton in vscode
}
\examples{
# ADD_EXAMPLES_HERE
}
