% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDateTz.R
\name{getDateTz}
\alias{getDateTz}
\title{Function to extract times from the GPIPP}
\usage{
getDateTz(x)
}
\arguments{
\item{x}{Vector full of date strings}
}
\description{
Converts to date and converts the more difficult strings in a second step if necessary
}
