% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giveCw.R
\name{giveCw}
\alias{giveCw}
\title{Crosswalk prepare function}
\usage{
giveCw(
  cw,
  from = NULL,
  to = NULL,
  by = NULL,
  sameLevel = TRUE,
  missings = c(" ", "99999", NA),
  ftb = c(from, to, by)
)
}
\arguments{
\item{cw}{A crosswalk as data.frame}

\item{from}{String indicating variable from which the crosswalk should happen}

\item{to}{String indicating variable to which crosswalk should happen}

\item{by}{String indicating variable for selecting which case in case of duplicates}

\item{sameLevel}{Boolean indicating whether duplicates for the to-variable should be avoided}

\item{missings}{Values that should be removed}

\item{ftb}{Variables can also be given as vector}
}
\description{
A function that prepares a crosswalk containing no missings and no duplicates based on a given variable
}
