% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipumsGenderAge.R
\name{ipumsGenderAge}
\alias{ipumsGenderAge}
\title{A function that calculates share female median and mean age for IPUMS data}
\usage{
ipumsGenderAge(df, code)
}
\arguments{
\item{df}{The data frame to calculate for}

\item{code}{The three letter code for age}
}
\description{
Calculates share female, mean and interpolated median age for IPUMS data sets based on frequency tables.
}
