% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_mult.R
\name{tab_mult}
\alias{tab_mult}
\title{Crea una tabla de multiplicar}
\usage{
tab_mult(n, simple = FALSE, bonita = TRUE)
}
\arguments{
\item{n}{(numeric) Hasta qué número se creará la tabla}

\item{simple}{(boolean) Solo resultado simple. Falso por defecto}

\item{bonita}{(boolean) Si la tabla que se imprime debe tener formato. Verdadero por defecto}
}
\value{
matrix
}
\description{
Crea una tabla de multiplicar
}
\examples{
tab_mult(10)
tab_mult(5, simple = TRUE)
}
